dev-requirements.txt
# Include base requirements
-r requirements.txt

# Note: setuptools>=67.0.0 is installed directly in setup.sh
# to ensure distutils is available for pandas-datareader

# Testing
pytest>=7.4,<8.0
pytest-cov>=4.1,<5.0

# Code quality tools
flake8>=6.1,<7.0
black>=23.12,<24.0
isort>=5.13,<6.0
pylint>=3.0,<4.0

# Type checking
mypy>=1.8,<2.0
types-requests>=2.31,<3.0
pandas-stubs>=2.1,<3.0
